+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 * Add `EveOnline::ESI::UniverseAsteroidBelt`
 * Add `EveOnline::ESI::UniverseMoon`
 * Add `EveOnline::ESI::Models::Moon`
+* Add `EveOnline::ESI::Models::SystemJump`
+* Add `EveOnline::ESI::UniverseSystemJumps`
+* Add `EveOnline::ESI::Models::SystemKill`
+* Add `EveOnline::ESI::UniverseSystemKills`
 
 ## v0.19.0